# config/documentation/swagger-admin.json
{
	"swagger": "2.0",
	"info": {
		"title": "HonorVet - Admin API",
		"description": "HonorVet - Admin API Documentation",
		"version": "1.0"
	},
	"produces": ["application/json"],
	"basePath": "/",
	"securityDefinitions": {
		"Bearer": {
			"type": "apiKey",
			"name": "authorization",
			"in": "header"
		}
	},
	"paths": {
...
		"/admin/v1/admin/": {
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-admin-put",
				"tags": ["Sub Admin"],
				"description": "Update Admin Details",
				"security": [
					{
						"Bearer": []
					}
				],
				"content": "multipart/form-data",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"properties": {
								"id": {
									"description": "id GUID",
									"required": false,
									"type": "string",
									"value": ""
								},
								"email": {
									"description": "email",
									"required": false,
									"type": "string",
									"value": ""
								},
								"firstName": {
									"description": "First Name",
									"required": false,
									"type": "string",
									"value": ""
								},
								"lastName": {
									"description": "Last Name",
									"type": "string",
									"value": ""
								},
								"adminType": {
									"enum": ["SUPER_ADMIN", "SUB_ADMIN"],
									"description": "",
									"type": "string",
									"required": false
								},
								"accessPermissions": {
									"type": "array",
									"required": false,
									"items": {
										"$ref": "#/definitions/accessPermissions"
									}
								}
							}
						}
					}
				],
				"responses": {}
			},
			"post": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-admin-post",
				"security": [
					{
						"Bearer": []
					}
				],
				"tags": ["Sub Admin"],
				"description": "Add SuperAdmins or SubAdmins",
				"consumes": "application/json",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"properties": {
								"email": {
									"description": "email is required",
									"required": true,
									"type": "string",
									"value": ""
								},
								"firstName": {
									"description": "First Name",
									"required": true,
									"type": "string",
									"value": ""
								},
								"lastName": {
									"description": "Last Name",
									"type": "string",
									"value": ""
								},
								"adminType": {
									"enum": ["SUPER_ADMIN", "SUB_ADMIN"],
									"description": "",
									"type": "string",
									"required": true
								},
								"accessPermissions": {
									"type": "array",
									"required": true,
									"items": {
										"$ref": "#/definitions/accessPermissions"
									}
								}
							}
						}
					}
				],
				"responses": {}
			},
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-admin-delete",
				"tags": ["Sub Admin"],
				"description": "Delete Admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/employmentTypes/": {
			"post": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminemploymentTypes-create-POST",
				"tags": ["Employment Types"],
				"description": "Add employmentTypes by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": true
					}
				],
				"responses": {}
			},
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminemploymentTypes-list-get",
				"tags": ["Employment Types"],
				"description": "Get employmentTypes by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			},
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminemploymentTypes-update-put",
				"tags": ["Employment Types"],
				"description": "Update employmentTypes by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					},
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": false
					},
					{
						"name": "isBlocked",
						"in": "formData",
						"description": "",
						"type": "number",
						"required": false
					}
				],
				"responses": {}
			},
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminemploymentTypes-delete-delete",
				"tags": ["Employment Types"],
				"description": "Delete employmentTypes by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/employmentTypes/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminemploymentTypes-detail-{id}-get",
				"tags": ["Employment Types"],
				"description": "Get single employmentTypes by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/benefits/": {
			"post": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminbenefits-create-POST",
				"tags": ["Benefits"],
				"description": "Add Benefits by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": true
					}
				],
				"responses": {}
			},
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminbenefits-list-get",
				"tags": ["Benefits"],
				"description": "Get benefits by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			},
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminBenefits-update-put",
				"tags": ["Benefits"],
				"description": "Update benefits by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					},
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": false
					},
					{
						"name": "isBlocked",
						"in": "formData",
						"description": "",
						"type": "number",
						"required": false
					}
				],
				"responses": {}
			},
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminbenefits-delete-delete",
				"tags": ["Benefits"],
				"description": "Delete benefits by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/benefits/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminbenefits-detail-{id}-get",
				"tags": ["Benefits"],
				"description": "Get single Benefits by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/hire/": {
			"post": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminHire-create-POST",
				"tags": ["Hire"],
				"description": "Add Hire by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": true
					}
				],
				"responses": {}
			},
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminHire-list-get",
				"tags": ["Hire"],
				"description": "Get Hire by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			},
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminHire-update-put",
				"tags": ["Hire"],
				"description": "Update Hire by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					},
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": false
					},
					{
						"name": "isBlocked",
						"in": "formData",
						"description": "",
						"type": "number",
						"required": false
					}
				],
				"responses": {}
			},
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminHire-delete-delete",
				"tags": ["Hire"],
				"description": "Delete Hire by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/hire/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-adminHire-detail-{id}-get",
				"tags": ["Hire"],
				"description": "Get single Hire by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/JobTitle/": {
			"post": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Job-Title-create-POST",
				"tags": ["Job Title"],
				"description": "Add Job Title by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": true
					}
				],
				"responses": {}
			},
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Job-Title-list-get",
				"tags": ["Job Title"],
				"description": "Get skills by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			},
			"put": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Job-Title-update-put",
				"tags": ["Job Title"],
				"description": "Update Job Title by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					},
					{
						"name": "name",
						"in": "formData",
						"description": "name is required",
						"type": "string",
						"value": "",
						"required": false
					},
					{
						"name": "isBlocked",
						"in": "formData",
						"description": "",
						"type": "number",
						"required": false
					}
				],
				"responses": {}
			},
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Job-Title-delete-delete",
				"tags": ["Job Title"],
				"description": "Delete JobTitle by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/JobTitle/detail/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-Job-Title-detail-{id}-get",
				"tags": ["Job Title"],
				"description": "Get Job Title skills by admin",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"format": "uuid",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
		"/admin/v1/jobs": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-user-jobs-get",
				"tags": ["Jobs Posted"],
				"description": "Get All jobs",
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"name": "recuiterId",
						"in": "query",
						"description": "",
						"type": "string",
						"required": false
					},
					{
						"name": "limit",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "skip",
						"in": "query",
						"description": "",
						"type": "number",
						"required": false
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "orderBy",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"value": ""
					},
					{
						"name": "search",
						"in": "query",
						"description": "",
						"required": false,
						"type": "string",
						"allowReserved": true,
						"value": ""
					}
				],
				"responses": {}
			}
		},
		"/admin/v1/jobs/{id}": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "admin-v1-common-{id}-get",
				"tags": ["Jobs Posted"],
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "Get Job details by id",

				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "",
						"type": "string",
						"required": true
					}
				],
				"responses": {}
			}
		},
...
	},
	"definitions": {
		"accessPermissions": {
			"type": "object",
			"properties": {
				"module": {
					"type": "string"
				},
				"permission": {
					"enum": [1, 0],
					"description": "0 or 1 to set permission",
					"type": "integer"
				}
			},
			"example": {
				"module": "dashboard",
				"permission": 1
			}
		}
	}
}